Refactor into independent files
