+++ webapp/src/components/rhs/rhs_about.tsx
@@ -10,11 +10,11 @@ import {setOwner, changeChannelName, updatePlaybookRunDescription} from 'src/cli
 import ProfileSelector from 'src/components/profile/profile_selector';
 import RHSPostUpdate from 'src/components/rhs/rhs_post_update';
 import {useProfilesInCurrentChannel} from 'src/hooks';
-import PostText from 'src/components/post_text';
 import RHSParticipants from 'src/components/rhs/rhs_participants';
 import {HoverMenu} from 'src/components/rhs/rhs_shared';
 import RHSAboutButtons from 'src/components/rhs/rhs_about_buttons';
-import {useClickOutsideRef, useKeyPress} from 'src/hooks/general';
+import RHSAboutTitle, {DefaultRenderedTitle} from 'src/components/rhs/rhs_about_title';
+import RHSAboutDescription from 'src/components/rhs/rhs_about_description';
 
 interface Props {
     playbookRun: PlaybookRun;
@@ -63,13 +63,14 @@ const RHSAbout = (props: Props) => {
                     toggleCollapsed={toggleCollapsed}
                 />
             </ButtonsRow>
-            <Title
+            <RHSAboutTitle
                 value={props.playbookRun.name}
                 onEdit={onTitleEdit}
+                renderedTitle={RenderedTitle}
             />
             {!collapsed &&
             <>
-                <Description
+                <RHSAboutDescription
                     value={props.playbookRun.description}
                     onEdit={onDescriptionEdit}
                 />
@@ -102,80 +103,6 @@ const RHSAbout = (props: Props) => {
         </Container>
     );
 };
-
-interface DescriptionProps {
-    value: string;
-    onEdit: (value: string) => void;
-}
-
-const Description = (props: DescriptionProps) => {
-    const placeholder = 'No description yet. Click here to edit it.';
-
-    const [editing, setEditing] = useState(false);
-    const [editedValue, setEditedValue] = useState(props.value || placeholder);
-
-    const textareaRef = useRef(null);
-
-    const saveAndClose = () => {
-        const newValue = editedValue.trim();
-        setEditedValue(newValue);
-        props.onEdit(newValue);
-        setEditing(false);
-    };
-
-    useClickOutsideRef(textareaRef, saveAndClose);
-    useKeyPress((e: KeyboardEvent) => e.ctrlKey && e.key === 'Enter', saveAndClose);
-
-    useEffect(() => {
-        setEditedValue(props.value || placeholder);
-    }, [props.value]);
-
-    if (!editing) {
-        return (
-            <RenderedDescription onClick={() => setEditing(true)}>
-                <PostText text={editedValue}/>
-            </RenderedDescription>
-        );
-    }
-
-    return (
-        <DescriptionTextArea
-            value={editedValue}
-            ref={textareaRef}
-            onChange={(e) => setEditedValue(e.target.value)}
-            autoFocus={true}
-            onFocus={(e) => {
-                const val = e.target.value;
-                e.target.value = '';
-                e.target.value = val;
-            }}
-            rows={editedValue.split('\n').length}
-        />
-    );
-};
-
-const DescriptionTextArea = styled.textarea`
-    resize: none;
-    width: 100%;
-    height: max-content;
-    padding: 4px 8px;
-    margin-top: -2px;
-    margin-bottom: 9px;
-
-    border: none;
-    border-radius: 5px;
-    box-shadow: none;
-
-    background: rgba(var(--center-channel-color-rgb), 0.04);
-
-    &:focus {
-        box-shadow: none;
-    }
-
-    font-size: 14px;
-    line-height: 15px;
-    color: var(--center-channel-color);
-`;
 
 const Container = styled.div`
     margin-top: 3px;
@@ -219,124 +146,18 @@ const ButtonsRow = styled(HoverMenu)`
     }
 `;
 
-const PaddedContent = styled.div`
-    padding: 0 8px; 
-`;
-
-interface TitleProps {
-    onEdit: (newTitle: string) => void;
-    value: string;
-}
-
-const Title = (props: TitleProps) => {
-    const [editing, setEditing] = useState(false);
-    const [editedValue, setEditedValue] = useState(props.value);
-
-    const inputRef = useRef(null);
-
-    useEffect(() => {
-        setEditedValue(props.value);
-    }, [props.value]);
-
-    const saveAndClose = () => {
-        props.onEdit(editedValue);
-        setEditing(false);
-    };
-
-    useClickOutsideRef(inputRef, saveAndClose);
-    useKeyPress('Enter', saveAndClose);
-
-    if (!editing) {
-        return (
-            <RenderedTitle onClick={() => setEditing(true)} >
-                {editedValue}
-            </RenderedTitle>
-        );
-    }
-
-    return (
-        <TitleInput
-            type={'text'}
-            ref={inputRef}
-            onChange={(e) => setEditedValue(e.target.value)}
-            value={editedValue}
-            maxLength={59}
-            autoFocus={true}
-            onFocus={(e) => {
-                const val = e.target.value;
-                e.target.value = '';
-                e.target.value = val;
-            }}
-        />
-    );
-};
-
-const TitleInput = styled.input`
-    width: calc(100% - 75px);
-    height: 30px;
-    padding: 4px 8px;
-    margin-bottom: 5px;
-    margin-top: -3px;
-
-    border: none;
-    border-radius: 5px;
-    box-shadow: none;
-
-    background: rgba(var(--center-channel-color-rgb), 0.04);
-
-    &:focus {
-        box-shadow: none;
-    }
-
-    color: var(--center-channel-color);
-    font-size: 18px;
-    line-height: 24px;
-    font-weight: 600;
-`;
-
-const RenderedTitle = styled(PaddedContent)`
-    max-width: 100%;
+const RenderedTitle = styled(DefaultRenderedTitle)`
     ${Container}:focus-within &, ${Container}:hover & {
         max-width: calc(100% - 75px);
     }
-
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-
-    height: 30px;
-    line-height: 24px;
-
-    font-size: 18px;
-    font-weight: 600;
-
-    color: var(--center-channel-color);
-
-    :hover {
-        cursor: text;
-    }
-
-    border-radius: 5px;
-
-    margin-bottom: 2px;
 `;
 
-const RenderedDescription = styled(PaddedContent)`
-    :hover {
-        cursor: text;
-    }
-
-    border-radius: 5px;
-
-    margin-bottom: 16px;
-    line-height: 20px;
-`;
-
-const Row = styled(PaddedContent)`
+const Row = styled.div`
     display: flex;
     flex-direction: row;
     flex-wrap: nowrap;
 
+    padding: 0 8px;
     margin-bottom: 30px;
 `;
 
@@ -359,11 +180,6 @@ const MemberSectionTitle = styled.div`
     line-height: 16px;
 
     color: rgba(var(--center-channel-color-rgb), 0.72)
-`;
-
-const NoDescription = styled.div`
-    color: rgba(var(--center-channel-color-rgb), 0.64);
-    margin-bottom: 10px;
 `;
 
 export default RHSAbout;
